# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText
  description: |-
    This OpenAPI document describes APIs for the project "WASAText".

    Student: Damir Tassybayev
    Matricola: 2046612
    E-mail: tassybayev.2046612@studenti.uniroma1.it
  version: "1.0.2"
...
security:
  - BearerAuth: []
paths:
  /session:
    post:
      tags:
        - login
      summary: Logs in the user
      description: |-
        If the user does not exist, it will be created and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      security: []
      requestBody:
        description: User details (name and photo in base64)
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                identifier: "abcdef012345"
...
  /groups:
    get:
      tags:
        - group
      summary: Retrieves the list of groups for the authenticated user
      description: Retrieves groups.
      operationId: getMyGroups
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of groups.
          content:
            application/json:
              schema:
                type: object
                description: Contains a list of groups.
                required:
                  - groups
                properties:
                  groups:
                    type: array
                    description: List of groups.
                    minItems: 0
                    maxItems: 1000
                    items:
                      $ref: '#/components/schemas/Group'
              example:
                groups:
                  - id: "group123"
                    name: "Group Chat"
                    members:
                      - "user123"
                      - "user456"
                    groupPhoto: "aGVsbG8="
    post:
      tags:
        - group
      summary: Creates a new group
      description: |-
        Creates a new group chat.
        Expects multipart/form-data with fields for group name, a JSON string of member IDs (in property "membersJson"),
        and a group image.
      operationId: createGroup
      security:
        - BearerAuth: []
      requestBody:
        description: Group creation data.
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: Form data for group creation.
              properties:
                name:
                  type: string
                  description: Name of the group.
                  example: "Group Chat"
                  pattern: '^[a-zA-Z0-9_ ]+$'
                  minLength: 3
                  maxLength: 50
                membersJson:
                  type: string
                  description: A JSON string array of member IDs (including the creator).
                  example: '["user123","user456"]'
                  pattern: '^\[.*\]$'
                  minLength: 2
                  maxLength: 1000
                image:
                  type: string
                  description: Group image (JPEG or PNG) as a base64 encoded string.
                  example: "aGVsbG8="
                  pattern: '^[A-Za-z0-9+/]*={0,2}$'
                  minLength: 1
                  maxLength: 10485760
      responses:
        '201':
          description: Group created successfully and returns the canonical Group object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
              example:
                id: "group123"
                name: "Group Chat"
                members:
                  - "user123"
                  - "user456"
                groupPhoto: "aGVsbG8="
  /groups/{groupId}:
    get:
      tags:
        - group
      summary: Retrieves details of a specific group
      description: Retrieves group details including members and group photo.
      operationId: getGroup
      security:
        - BearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          description: ID of the group.
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_]+$'
            minLength: 1
            maxLength: 50
      responses:
        '200':
          description: Group details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
              example:
                id: "group123"
                name: "Group Chat"
                members:
                  - "user123"
                  - "user456"
                groupPhoto: "aGVsbG8="
    delete:
      tags:
        - group
      summary: Leaves a group
      description: Removes the authenticated user from the specified group.
      operationId: leaveGroup
      security:
        - BearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          description: ID of the group to leave.
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_]+$'
            minLength: 1
            maxLength: 50
      responses:
        '204':
          description: Left group successfully.
    post:
      tags:
        - group
      summary: Adds a user to a group
      description: Adds a user to the specified group.
      operationId: addToGroup
      security:
        - BearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          description: ID of the group.
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_]+$'
            minLength: 1
            maxLength: 50
      requestBody:
        description: JSON payload with the user ID to add.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddGroupMemberRequest'
      responses:
        '204':
          description: User added to group successfully.
  /groups/{groupId}/name:
    put:
      tags:
        - group
      summary: Updates the group's name
      description: Updates the group's name.
      operationId: setGroupName
      security:
        - BearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          description: ID of the group.
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_]+$'
            minLength: 1
            maxLength: 50
      requestBody:
        description: New group name.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupRequest'
      responses:
        '200':
          description: Group name updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
              example:
                id: "group123"
                name: "New Group Name"
                members:
                  - "user123"
                  - "user456"
                groupPhoto: "aGVsbG8="
  /groups/{groupId}/photo:
    put:
      tags:
        - group
      summary: Updates the group's photo
      description: Updates the group's photo.
      operationId: setGroupPhoto
      security:
        - BearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          description: ID of the group.
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_]+$'
            minLength: 1
            maxLength: 50
      requestBody:
        description: New photo (JPEG or PNG) as a base64 encoded string.
        required: true
        content:
          image/png:
            schema:
              type: string
              description: PNG image file (base64 encoded)
              pattern: '^[A-Za-z0-9+/]*={0,2}$'
              minLength: 1
              maxLength: 10485760
          image/jpeg:
            schema:
              type: string
              description: JPEG image file (base64 encoded)
              pattern: '^[A-Za-z0-9+/]*={0,2}$'
              minLength: 1
              maxLength: 10485760
      responses:
        '200':
          description: Group photo updated successfully.
          content:
            application/json:
              schema:
                type: object
                description: Confirmation message for group photo update.
                required:
                  - message
                properties:
                  message:
                    type: string
                    description: Confirmation message.
                    example: "Group photo updated successfully"
                    pattern: '^.*$'
                    minLength: 1
                    maxLength: 100
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      description: Request schema for user login.
      required:
        - name
        - photo
      properties:
        name:
          type: string
          description: The name of the user to log in.
          example: "Maria"
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 3
          maxLength: 16
        photo:
          type: string
          description: Base64-encoded default photo.
          example: "aGVsbG8="
          pattern: '^[A-Za-z0-9+/]+={0,2}$'
          minLength: 1
          maxLength: 1000000
    LoginResponse:
      type: object
      description: Response schema for user login.
      properties:
        identifier:
          type: string
          description: The identifier of the logged-in user.
          example: "abcdef012345"
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 1
          maxLength: 50
...
    AddGroupMemberRequest:
      type: object
      description: Request schema for adding a group member.
      required:
        - userId
      properties:
        userId:
          type: string
          description: ID of the user to add.
          example: "user789"
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 1
          maxLength: 50
    UpdateGroupRequest:
      type: object
      description: Request schema for updating group information.
      required:
        - groupName
      properties:
        groupName:
          type: string
          description: New name of the group.
          example: "New Group Name"
          pattern: '^[a-zA-Z0-9_ ]+$'
          minLength: 3
          maxLength: 16
    Group:
      type: object
      description: Group schema.
      required:
        - id
        - name
        - members
      properties:
        id:
          type: string
          description: Unique identifier of the group.
          example: "group123"
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 1
          maxLength: 50
        name:
          type: string
          description: Name of the group.
          example: "Group Chat"
          pattern: '^[a-zA-Z0-9_ ]+$'
          minLength: 3
          maxLength: 50
        members:
          type: array
          description: List of user IDs that are members of the group.
          minItems: 1
          maxItems: 1000
          items:
            type: string
            description: User ID.
            example: "user123"
            pattern: '^[a-zA-Z0-9_]+$'
            minLength: 1
            maxLength: 50
        groupPhoto:
          type: string
          description: Base64-encoded group photo (if any).
          example: "aGVsbG8="
          pattern: '^[A-Za-z0-9+/]*={0,2}$'
          minLength: 0
          maxLength: 1000000